+++ src/main/java/com/sns/post/bo/PostBO.java
@@ -1,17 +1,15 @@
 package com.sns.post.bo;
 
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.log;
-
 import java.util.List;
-import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.sns.comment.bo.CommentBO;
 import com.sns.common.FileManagerService;
+import com.sns.like.bo.LikeBO;
 import com.sns.post.entity.PostEntity;
-import com.sns.post.mapper.PostMapper;
 import com.sns.post.repostiory.PostRepository;
 
 import lombok.extern.slf4j.Slf4j;
@@ -22,9 +20,15 @@ public class PostBO {
 	
 	@Autowired
 	private PostRepository postRepository;
-	
+
 	@Autowired
 	private FileManagerService fileManagerService;
+	
+	@Autowired
+	private CommentBO commentBO;
+	
+	@Autowired
+	private LikeBO likeBO;
 	
 	// input: X   output: List<PostEntity>
 	public List<PostEntity> getPostList() {
@@ -53,19 +57,23 @@ public class PostBO {
 		public void deletePostByPostIdUserId(int postId, int userId) {
 			
 			// 기존 글 가져오기
-			List<PostEntity> post = postRepository.findAllByOrderByIdDesc();
-			
+			PostEntity post = postRepository.findById(postId).orElse(null);
+			if (post == null) {
+				log.error("[delete post] postId:{}, userId:{}", postId, userId);
+				return;
+			}
+
 			// 글 삭제
 			postRepository.delete(post);
-			
+
 			// 이미지 있으면 삭제
 			fileManagerService.deleteFile(post.getImagePath());
-			
+
 			// 댓글들 삭제
-			fileManagerService.deleteFile(comment.content());
-			
+			commentBO.deleteCommentsByPostId(postId);
+
 			// 좋아요들 삭제
-			fileManagerService.deleteFile(like.postId());
+			likeBO.deleteLikeByPostId(postId);
 		}
 	
 }
